pyproject.toml
[tool.poetry]
name = "api-to-database"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "api_to_database"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.1"
requests = "^2.28.1"
black = "^22.10.0"
pytest = "^7.2.0"
sqlalchemy-orm = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
